requirements.txt
streamlit==1.35.0
langchain==0.2.0
langchain-community==0.2.0
langchain-core==0.2.0
langchain-groq==0.1.4
langchain-huggingface==0.0.3
langchain-text-splitters==0.2.0
chromadb==0.5.0
flashrank==0.2.0
pymupdf==1.24.4
python-dotenv==1.0.1
numpy<2.0.0